calculation fix

+++ src/main/java/com/rotterdam/model/entity/WorkHour.java
@@ -38,7 +38,7 @@ public class WorkHour implements HibernateL2Cache, Serializable {
     public double calculateTotal(){
         double endTime = DateTools.getDoubleFormatHours(endWorkingTime);
         double startTime = DateTools.getDoubleFormatHours(startWorkingTime);
-        int rT = restTime / 60;
+        double rT = (double)restTime / 60;
         return endTime - startTime - rT;
     }
